+++ pynq/lib/arduino/arduino_grove_gesture/Debug/objects.mk
@@ -1,0 +1,8 @@
+################################################################################
+# Automatically-generated file. Do not edit!
+################################################################################
+
+USER_OBJS :=
+
+LIBS := -Wl,--start-group,-lxil,-lgcc,-lc,--end-group
+
