+++ pynq/lib/arduino/arduino_grove_gesture/Debug/sources.mk
@@ -1,0 +1,20 @@
+################################################################################
+# Automatically-generated file. Do not edit!
+################################################################################
+
+LD_SRCS := 
+OBJ_SRCS := 
+S_SRCS := 
+C_SRCS := 
+S_UPPER_SRCS := 
+O_SRCS := 
+EXECUTABLES := 
+OBJS := 
+S_UPPER_DEPS := 
+C_DEPS := 
+ELFSIZE := 
+
+# Every subdirectory with source files must be described here
+SUBDIRS := \
+src \
+
