+++ pynq/lib/arduino/arduino_grove_gesture/Debug/src/subdir.mk
@@ -1,0 +1,30 @@
+################################################################################
+# Automatically-generated file. Do not edit!
+################################################################################
+
+# Add inputs and outputs from these tool invocations to the build variables 
+LD_SRCS += \
+../src/lscript.ld 
+
+C_SRCS += \
+../src/arduino_grove_gesture.c \
+../src/grove_gesture.c 
+
+OBJS += \
+./src/arduino_grove_gesture.o \
+./src/grove_gesture.o 
+
+C_DEPS += \
+./src/arduino_grove_gesture.d \
+./src/grove_gesture.d 
+
+
+# Each subdirectory must supply rules for building sources it contributes
+src/%.o: ../src/%.c
+	@echo 'Building file: $<'
+	@echo 'Invoking: MicroBlaze gcc compiler'
+	mb-gcc -Wall -O3 -g3 -c -fmessage-length=0 -MT"$@" -I../../bsp_iop_arduino/iop_arduino_mb/include -mlittle-endian -mcpu=v10.0 -mxl-soft-mul -Wl,--no-relax -MMD -MP -MF"$(@:%.o=%.d)" -MT"$(@:%.o=%.d)" -o "$@" "$<"
+	@echo 'Finished building: $<'
+	@echo ' '
+
+
